SQuAD 2.0 compatibility
