+++ allennlp_models/rc/metrics/__init__.py
@@ -1,3 +1,2 @@
 from allennlp_models.rc.metrics.drop_em_and_f1 import DropEmAndF1
 from allennlp_models.rc.metrics.squad_em_and_f1 import SquadEmAndF1
-from allennlp_models.rc.metrics.squad2_em_and_f1 import Squad2EmAndF1
